Extract Italian Michela and fix Japanese export
